연산자

+++ 221226/hw1.c
@@ -3,9 +3,9 @@
 
 int main() {
 
-    double r;
+    int r;
     printf("원의 반지름을 입력하세요.\n반지름 : ");
-    scanf("%lf", &r);
+    scanf("%d", &r);
     printf("원의 넓이는 %.2lf입니다.\n", r * r * PI);
 
     return 0;
